mangatown indexer updated

+++ model/chapter.go
@@ -11,7 +11,7 @@ type Chapter struct {
 	CreatedAt time.Time          `json:"createdAt" bson:"createdAt"`
 	UpdatedAt time.Time          `json:"updatedAt" bson:"updatedAt"`
 	Title     *string            `json:"title" bson:"title"`
-	Number    int64              `json:"number" bson:"number"`
+	Number    float64            `json:"number" bson:"number"`
 	Links     []string           `json:"links" bson:"links"`
 	Source    string             `json:"source" bson:"source"`
 	Language  string             `json:"language" bson:"language"`
